requirements.txt
pytesseract
pillow
openai
pyautogui
Opencv-python